Add client models for locations. Convert Trips and Activities to use location classes. Add placeManager and locationBuilder classes to help with constructing locations.

+++ server/src/main/java/com/tripdazzle/server/datamodels/location/CustomLocationData.java
@@ -2,10 +2,10 @@ package com.tripdazzle.server.datamodels.location;
 
 public class CustomLocationData implements LocationData {
     public final String name;
-    public final float latitude;
-    public final float longitude;
+    public final double latitude;
+    public final double longitude;
 
-    public CustomLocationData(String name, float latitude, float longitude) {
+    public CustomLocationData(String name, double latitude, double longitude) {
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
